Allow an offset when calculating an EMA
When creating a MACD signal line, using the 9 period EMA results in a
long period of time (~60 or so candles) before the macdhistogram
accurately reflects momentum. This is because from periods 12 - 26
(assuming the short window is 12 and long window 26) are just an EMA of
the short window.

Creating an EMA with offset allows us to skip calculating the EMA during
those periods, resulting in quicker accuracy of the macdhistogram.